fix: module exposure
