+++ marimo/_runtime/watch/__init__.py
@@ -1,5 +1,9 @@
 # Copyright 2025 Marimo. All rights reserved.
-from marimo._runtime.watch.watch import directory, file
+from marimo._runtime.watch._directory import directory
+from marimo._runtime.watch._file import file
+
+# NB. _runtime/reload captures module level changes and
+# marimo/_server/sessions.py captures notebook level changes.
 
 __all__ = [
     "file",
