--- marimo/_runtime/watch/watch.py
@@ -1,341 +1,0 @@
-# Copyright 2024 Marimo. All rights reserved.
-from __future__ import annotations
-
-import os
-import sys
-import threading
-import time
-from pathlib import Path
-from typing import TYPE_CHECKING, Any, Callable, TypeVar, cast
-
-from marimo._output.rich_help import mddoc
-from marimo._runtime.context import (
-    ContextNotInitializedError,
-    get_context,
-    runtime_context_installed,
-)
-from marimo._runtime.side_effect import SideEffect
-from marimo._runtime.state import State
-from marimo._runtime.threads import Thread
-
-if TYPE_CHECKING:
-    from collections.abc import Iterable
-
-if sys.version_info >= (3, 11):
-    from typing import Self
-else:
-    from typing_extensions import Self
-
-
-T = TypeVar("T")
-
-WATCHER_SLEEP_INTERVAL = 1.0
-
-# For testing only - do not use in production
-_TEST_SLEEP_INTERVAL: float | None = None
-
-
-def write_side_effect(data: str | bytes) -> None:
-    """Write side effect to the context."""
-    try:
-        ctx = get_context()
-    except ContextNotInitializedError:
-        # Context is not initialized, nothing we can do
-        return
-    ctx.cell_lifecycle_registry.add(SideEffect(data))
-
-
-def watch_file(
-    path: Path, state: FileState, should_exit: threading.Event
-) -> None:
-    """Watch a file for changes and update the state."""
-    last_mtime: float = 0
-    current_mtime = last_mtime
-    sleep_interval = _TEST_SLEEP_INTERVAL or WATCHER_SLEEP_INTERVAL
-    while not should_exit.is_set():
-        time.sleep(sleep_interval)
-        try:
-            current_mtime = path.stat().st_mtime
-        except FileNotFoundError:
-            # File has been deleted, trigger a change
-            current_mtime = 0
-        except Exception as e:
-            # Handle other exceptions (e.g., permission denied)
-            sys.stderr.write(f"Error watching file {path}: {e}\n")
-            continue
-
-        if current_mtime != last_mtime:
-            last_mtime = current_mtime
-            state._set_value(path)
-
-
-def walk(path: Path) -> Iterable[tuple[Path, list[str], list[str]]]:
-    if sys.version_info > (3, 11):
-        return path.walk()
-    return os.walk(path)
-
-
-def _hashable_walk(
-    walked: Iterable[tuple[Path, list[str], list[str]]],
-) -> set[tuple[Path, tuple[str], tuple[str]]]:
-    return cast(
-        set[tuple[Path, tuple[str], tuple[str]]],
-        set((p, *map(tuple, r)) for p, *r in walked),
-    )
-
-
-def hashable_walk(path: Path) -> set[tuple[Path, tuple[str], tuple[str]]]:
-    return _hashable_walk(walk(path))
-
-
-def watch_directory(
-    path: Path, state: DirectoryState, should_exit: threading.Event
-) -> None:
-    """Watch a directory for changes and update the state."""
-    last_structure = hashable_walk(path)
-    current_structure = last_structure
-    sleep_interval = _TEST_SLEEP_INTERVAL or WATCHER_SLEEP_INTERVAL
-    while not should_exit.is_set():
-        time.sleep(sleep_interval)
-        try:
-            current_structure = hashable_walk(path)
-        except FileNotFoundError:
-            # Directory has been deleted, trigger a change
-            current_structure = set()
-        except Exception as e:
-            # Handle other exceptions (e.g., permission denied)
-            sys.stderr.write(f"Error watching directory {path}: {e}\n")
-            continue
-
-        if current_structure != last_structure:
-            last_structure = current_structure
-            state._set_value(path)
-
-
-class PathState(State[Path]):
-    """Base class for path state."""
-
-    _forbidden_attributes: set[str]
-    _target: Callable[[Path, Self, threading.Event], None]
-
-    def __init__(
-        self,
-        path: Path,
-        *args: Any,
-        allow_self_loops: bool = True,
-        **kwargs: Any,
-    ) -> None:
-        if kwargs.pop("_context", None) is not None:
-            raise ValueError(
-                "The '_context' argument is not supported for this class."
-            )
-        if kwargs.pop("allow_self_loops", None) is not None:
-            raise ValueError(
-                "The 'allow_self_loops' argument is not supported for this class."
-            )
-
-        # Mypy seems to think we could provide multiple kwargs definitions here
-        # but we can't.
-        super().__init__(
-            path,
-            *args,
-            allow_self_loops=allow_self_loops,
-            _context="file",
-            **kwargs,
-        )  # type: ignore[misc]
-        self._should_exit = threading.Event()
-        # Only bother with the watcher if the context is installed
-        # State is not enabled in script mode
-        if runtime_context_installed():
-            Thread(
-                target=self._target,
-                args=(path, self, self._should_exit),
-                daemon=True,
-            ).start()
-
-    def __getattr__(self, name: str) -> Any:
-        """Get an attribute from the file path."""
-        # Disable some attributes
-        if name in self._forbidden_attributes:
-            raise AttributeError(
-                f"'{self.__class__.__name__}' does not "
-                f"expose attribute '{name}'"
-            )
-        if hasattr(self._value, name):
-            return getattr(self._value, name)
-        raise AttributeError(
-            f"'{self.__class__.__name__}' object has no attribute '{name}'"
-        )
-
-    def __del__(self) -> None:
-        self._should_exit.set()
-
-    def __repr__(self) -> str:
-        return f"{self.__class__.__name__}({self._value})"
-
-    def exists(self) -> bool:
-        """Check if the path exists."""
-        exists = self._value.exists()
-        if not exists:
-            write_side_effect(f"doesn't exists:{self._value}")
-        else:
-            _ = self.read_text()
-        return exists
-
-
-class FileState(PathState):
-    """Wrapper for file state."""
-
-    _forbidden_attributes = {
-        "open",
-        "iterdir",
-        "glob",
-        "rglob",
-        "mkdir",
-        "rename",
-        "replace",
-        "walk",
-    }
-    _target: Callable[[Path, FileState, threading.Event], None] = staticmethod(
-        watch_file
-    )
-
-    def read_text(self) -> str:
-        """Read the file as a string."""
-        text = self._value.read_text()
-        write_side_effect(f"read_text:{text}")
-        return text
-
-    def write_text(self, value: str) -> int:
-        """Write the file as a string."""
-        response = self._value.write_text(value)
-        text = self._value.read_text()
-        write_side_effect(f"write_text:{text}")
-        return response
-
-    def read_bytes(self) -> bytes:
-        """Read the file as bytes."""
-        data = self._value.read_bytes()
-        write_side_effect(f"read_bytes:{data!r}")
-        return data
-
-    def write_bytes(self, value: bytes) -> int:
-        """Write the file as bytes."""
-        response = self._value.write_bytes(value)
-        data = self._value.read_bytes()
-        write_side_effect(f"write_bytes:{data!r}")
-        return response
-
-
-class DirectoryState(PathState):
-    """Wrapper for directory state."""
-
-    _forbidden_attributes = {
-        "open",
-        "rename",
-        "replace",
-        "remove",
-        "unlink",
-        "write_text",
-        "write_bytes",
-        "read_text",
-        "read_bytes",
-        "mkdir",
-        "touch",
-    }
-    _target: Callable[[Path, DirectoryState, threading.Event], None] = (
-        staticmethod(watch_directory)
-    )
-
-    def walk(self) -> Iterable[tuple[Path, list[str], list[str]]]:
-        """Walk the directory."""
-        items = walk(self._value)
-        as_list = list(_hashable_walk(items))
-        write_side_effect(f"walk:{sorted(as_list)}")
-        return iter(items)
-
-    def iterdir(self) -> Iterable[Path]:
-        """Iterate over the directory."""
-        items = list(self._value.iterdir())
-        write_side_effect(f"iterdir:{items}")
-        return iter(items)
-
-    def glob(self, pattern: str) -> Iterable[Path]:
-        """Glob the directory."""
-        items = list(self._value.glob(pattern))
-        write_side_effect(f"glob:{items}")
-        return iter(items)
-
-    def rglob(self, pattern: str) -> Iterable[Path]:
-        """Recursive glob the directory."""
-        items = list(self._value.rglob(pattern))
-        write_side_effect(f"rglob:{items}")
-        return iter(items)
-
-
-@mddoc
-def file(path: Path | str) -> FileState:
-    """
-    A reactive wrapper for file paths.
-
-    This function takes a file path to watch and returns a wrapper to reactively
-    read and write from the file.
-
-    The "wrapped" file Path object exposes most of the same methods as the
-    [pathlib.Path object](https://docs.python.org/3/library/pathlib.html#pathlib.Path),
-    with a few exceptions. The following methods are not available:
-    - `open()`
-    - `rename()`
-    - `replace()`
-
-    This object will trigger dependent cells to re-evaluate when the file is
-    changed.
-
-    Warning:
-        It is possible to misuse this API in similar ways to `state()`. Consider
-        reading the warning and caveats in the `state()` documentation, and
-        using this function only when reading file paths, and not when writing
-        them.
-
-    Args:
-      path: Path to watch.
-
-    Returns:
-        A reactive wrapper for watching the file path.
-    """
-    if isinstance(path, str):
-        path = Path(path)
-    if path.is_dir():
-        raise ValueError(
-            "Path must be a file, not a directory, use mo.directory() instead"
-        )
-    return FileState(path, allow_self_loops=True)
-
-
-@mddoc
-def directory(path: Path | str) -> DirectoryState:
-    """
-    A reactive wrapper for directory paths.
-
-    This function takes a directory path to watch and returns a wrapper to
-    reactively list the contents of the directory.
-
-    This object will trigger dependent cells to re-evaluate when the directory
-    structure is changed (i.e., files are added or removed).
-
-    NOTE: This function does NOT react to file content changes, only to changes
-    in the directory structure. Utilize `mo.file()` to watch for changes in
-    specific files. Additional note: this will not follow symlinks.
-
-    Args:
-      path: Path to watch.
-
-    Returns:
-        A reactive wrapper for watching the directory.
-    """
-
-    if isinstance(path, str):
-        path = Path(path)
-    if not path.is_dir():
-        raise ValueError("Path must be a directory, use mo.file()")
-    return DirectoryState(path, allow_self_loops=True)
